trying to get add hunt to work
