+++ client/src/app/add-hunt/add-hunt.component.html
@@ -1,6 +1,6 @@
 <div class="flex-row">
   <div class="flex-1">
-    <form [formGroup]="addUserForm" (ngSubmit)="submitForm();">
+    <form [formGroup]="addHuntForm" (ngSubmit)="submitForm();">
       <mat-card class="add-user-card">
         <mat-card-header>
           <mat-card-title class="add-user-title">New User</mat-card-title>
@@ -8,17 +8,17 @@
         <mat-card-content>
           <div class="flex-col">
             <mat-form-field>
-              <mat-label>Name</mat-label>
-              <input matInput formControlName="name"
+              <mat-label>Title</mat-label>
+              <input matInput formControlName="title"
                 required>
-              @if (formControlHasError('name')) {
-                <mat-error data-test="nameError">
-                  {{getErrorMessage('name')}}
+              @if (formControlHasError('title')) {
+                <mat-error data-test="title">
+                  {{getErrorMessage('title')}}
                 </mat-error>
               }
             </mat-form-field>
 
-            <mat-form-field>
+            <!-- <mat-form-field>
               <mat-label>Age</mat-label>
               <input matInput type="number"
                 formControlName="age" required>
@@ -27,15 +27,15 @@
                   {{getErrorMessage('age')}}
                 </mat-error>
               }
-            </mat-form-field>
+            </mat-form-field> -->
 
-            <mat-form-field>
+            <!-- <mat-form-field>
               <mat-label>Company</mat-label>
               <input matInput formControlName="company">
               <mat-hint>Not required in Angular for demonstration, still required by server</mat-hint>
-            </mat-form-field>
+            </mat-form-field> -->
 
-            <mat-form-field>
+            <!-- <mat-form-field>
               <mat-label>Email</mat-label>
               <input matInput type="email"
                 formControlName="email" required>
@@ -44,9 +44,9 @@
                   {{getErrorMessage('email')}}
                 </mat-error>
               }
-            </mat-form-field>
+            </mat-form-field> -->
 
-            <mat-form-field>
+            <!-- <mat-form-field>
               <mat-label>Role</mat-label>
               <mat-select required formControlName="role">
                 <mat-option value="admin">Admin</mat-option>
@@ -58,7 +58,7 @@
                   {{getErrorMessage('role')}}
                 </mat-error>
               }
-            </mat-form-field>
+            </mat-form-field> -->
           </div>
 
         </mat-card-content>
@@ -67,9 +67,9 @@
             mat-button
             type="button"
             color="primary"
-            [disabled]="!addUserForm.valid"
+            [disabled]="!addHuntForm.valid"
             type="submit"
-            data-test="confirmAddUserButton"
+            data-test="confirmAddHuntButton"
             >
             ADD USER
           </button>
