+++ client/src/app/app-routing.module.ts
@@ -5,10 +5,10 @@ import { AddUserComponent } from './users/add-user.component';
 import { UserListComponent } from './users/user-list.component';
 import { UserProfileComponent } from './users/user-profile.component';
 import { CompanyListComponent } from './company-list/company-list.component';
-import { HostComponent } from './host/host.component';
 import { PlayerComponent } from './player/player.component';
 import { HuntListComponent } from './hunts/hunt-list.component';
 import { HttpClientModule } from '@angular/common/http';
+import { AddHuntComponent } from './add-hunt/add-hunt.component';
 
 
 // Note that the 'users/new' route needs to come before 'users/:id'.
@@ -20,9 +20,10 @@ const routes: Routes = [
   {path: 'users/new', component: AddUserComponent, title: 'Add User'},
   {path: 'users/:id', component: UserProfileComponent, title: 'User Profile'},
   {path: 'companies', component: CompanyListComponent, title: 'Companies'},
-  {path: 'host', component: HostComponent, title: 'Host'},
   {path: 'player', component: PlayerComponent, title: 'Player'},
-  {path: 'hunts', component: HuntListComponent, title: 'Hunts'}
+  {path: 'hunts', component: HuntListComponent, title: 'Hunts'},
+  {path: 'add-hunt', component: AddHuntComponent, title: 'Hunts'},
+
 
 ];
