+++ client/src/app/app.component.html
@@ -11,7 +11,7 @@
         Home
       </a>
 
-      <a mat-list-item routerLink="/host" routerLinkActive="drawer-list-item-active" (click)="drawer.close()">
+      <a mat-list-item routerLink="/add-hunt" routerLinkActive="drawer-list-item-active" (click)="drawer.close()">
         <mat-icon matListItemIcon>people</mat-icon>
         Host
       </a>
