+++ client/src/app/home/home.component.html
@@ -6,7 +6,7 @@
         <div class="card-body">
           <h1 class="card-title">Start New Hunt:</h1>
           <h3 class="card-text">Create a Scavenger Hunt as a host -</h3>
-          <a routerLink="/start-new-hunt-component" href="host" class="btn btn-primary btn-lg">Click Here</a>
+          <a routerLink="/start-new-hunt-component" href="add-hunt" class="btn btn-primary btn-lg">Click Here</a>
         </div>
       </div>
     </div>
