--- client/src/app/host/host.component.html
@@ -1,80 +1,0 @@
-<div class="flex-row">
-  <div class="flex-1">
-    <form [formGroup]="addUserForm" (ngSubmit)="submitForm();">
-      <mat-card class="add-user-card">
-        <mat-card-header>
-          <mat-card-title class="add-user-title">New User</mat-card-title>
-        </mat-card-header>
-        <mat-card-content>
-          <div class="flex-col">
-            <mat-form-field>
-              <mat-label>Name</mat-label>
-              <input matInput formControlName="name"
-                required>
-              @if (formControlHasError('name')) {
-                <mat-error data-test="nameError">
-                  {{getErrorMessage('name')}}
-                </mat-error>
-              }
-            </mat-form-field>
-
-            <mat-form-field>
-              <mat-label>Age</mat-label>
-              <input matInput type="number"
-                formControlName="age" required>
-              @if (formControlHasError('age')) {
-                <mat-error data-test="ageError">
-                  {{getErrorMessage('age')}}
-                </mat-error>
-              }
-            </mat-form-field>
-<!--
-            <mat-form-field>
-              <mat-label>Company</mat-label>
-              <input matInput formControlName="company">
-              <mat-hint>Not required in Angular for demonstration, still required by server</mat-hint>
-            </mat-form-field>
-
-            <mat-form-field>
-              <mat-label>Email</mat-label>
-              <input matInput type="email"
-                formControlName="email" required>
-              @if (formControlHasError('email')) {
-                <mat-error data-test="emailError">
-                  {{getErrorMessage('email')}}
-                </mat-error>
-              }
-            </mat-form-field> -->
-
-            <!-- <mat-form-field>
-              <mat-label>Role</mat-label>
-              <mat-select required formControlName="role">
-                <mat-option value="admin">Admin</mat-option>
-                <mat-option value="editor">Editor</mat-option>
-                <mat-option value="viewer">Viewer</mat-option>
-              </mat-select>
-              @if (formControlHasError('role')) {
-                <mat-error data-test="roleError">
-                  {{getErrorMessage('role')}}
-                </mat-error>
-              }
-            </mat-form-field> -->
-          </div>
-
-        </mat-card-content>
-        <mat-card-actions align="end">
-          <button
-            mat-button
-            type="button"
-            color="primary"
-            [disabled]="!addUserForm.valid"
-            type="submit"
-            data-test="confirmAddUserButton"
-            >
-            ADD USER
-          </button>
-        </mat-card-actions>
-      </mat-card>
-    </form>
-  </div>
-</div>
