--- client/src/app/hunts/add-hunt.component.ts
@@ -1,23 +1,0 @@
-import { Component } from '@angular/core';
-import { FormControl, FormGroup, FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
-import { MatButtonModule } from '@angular/material/button';
-import { MatCardModule } from '@angular/material/card';
-import { MatOptionModule } from '@angular/material/core';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatInputModule } from '@angular/material/input';
-import { MatSelectModule } from '@angular/material/select';
-import { MatSnackBar } from '@angular/material/snack-bar';
-import { Router } from '@angular/router';
-import { Hunt } from './hunt';
-import { HuntService } from './hunt.service';
-
-@Component({
-  selector: 'app-add-hunt',
-  standalone: true,
-  imports: [FormsModule, ReactiveFormsModule, MatCardModule, MatFormFieldModule, MatInputModule, MatSelectModule, MatOptionModule, MatButtonModule],
-  templateUrl: './add-hunt.component.html',
-  styleUrl: './add-hunt.component.scss'
-})
-export class AddHuntComponent {
-
-}
