+++ client/src/app/hunts/hunt.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
+import { Observable, map } from 'rxjs';
 import { environment } from '../../environments/environment';
 import { Hunt } from './hunt';
 
@@ -13,6 +13,7 @@ export class HuntService {
   readonly huntUrl: string = `${environment.apiUrl}hunts`;
 
   private readonly hostKey = 'hostid';
+  private readonly titleKey = 'title';
 
   constructor(private httpClient: HttpClient) {
   }
@@ -25,17 +26,26 @@ export class HuntService {
   *   from the server after a possibly substantial delay (because we're
   *   contacting a remote server over the Internet).
   */
-  getHunts(filters?: { hostid?: string }): Observable<Hunt[]> {
+  getHunts(filters?: { hostid?: string; title?: string}): Observable<Hunt[]> {
     let httpParams: HttpParams = new HttpParams();
     if (filters) {
       if (filters.hostid) {
         httpParams = httpParams.set(this.hostKey, filters.hostid);
       }
+      if (filters.title) {
+        httpParams = httpParams.set(this.titleKey, filters.title);
+      }
+
     }
 
     return this.httpClient.get<Hunt[]>(this.huntUrl, {
       params: httpParams
     });
+  }
+
+  addHunt(newHunt: Partial<Hunt>): Observable<string> {
+    // Send post request to add a new user with the user data as the body.
+    return this.httpClient.post<{id: string}>(this.huntUrl, newHunt).pipe(map(res => res.id));
   }
 }
